+++ .npmrc
@@ -1,0 +1,2 @@
+hoist-pattern[]=!@react-native
+hoist-pattern[]=!metro-config
