+++ apps/app/__tests__/App.test.tsx
@@ -4,7 +4,7 @@
 
 import 'react-native'
 import React from 'react'
-import App from '../App'
+import App from '../src/App.tsx'
 
 // Note: import explicitly to use the types shipped with jest.
 import { it } from '@jest/globals'
