+++ apps/app/android/settings.gradle
@@ -1,6 +1,8 @@
 pluginManagement { includeBuild("../node_modules/@react-native/gradle-plugin") }
 plugins { id("com.facebook.react.settings") }
 extensions.configure(com.facebook.react.ReactSettingsExtension){ ex -> ex.autolinkLibrariesFromCommand() }
-rootProject.name = 'app'
+rootProject.name = 'fienmee-app'
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
+android.useAndroidX=true
+android.enableJetifier=true
