+++ apps/app/package.json
@@ -10,8 +10,10 @@
         "test": "jest"
     },
     "dependencies": {
+        "metro-config": "^0.81.0",
         "react": "18.3.1",
-        "react-native": "0.76.2"
+        "react-native": "0.76.2",
+        "react-native-webview": "^13.12.4"
     },
     "devDependencies": {
         "@babel/core": "^7.25.2",
@@ -21,7 +23,9 @@
         "@react-native-community/cli-platform-android": "15.0.1",
         "@react-native-community/cli-platform-ios": "15.0.1",
         "@react-native/babel-preset": "0.76.2",
+        "@react-native/codegen": "^0.76.3",
         "@react-native/eslint-config": "0.76.2",
+        "@react-native/gradle-plugin": "^0.76.2",
         "@react-native/metro-config": "0.76.2",
         "@react-native/typescript-config": "0.76.2",
         "@types/react": "^18.2.6",
